requirements.txt
torch
numpy>=1.17.4
h5py>=3.3.0
hdf5plugin>=3.2.0
tables>=3.6.1
pandas>=1.1.5
Pillow-SIMD
torchvision
scikit-image
psutil
tqdm
pyturbojpeg
cffi
jpegtran-cffi
simplejpeg


